requirements.txt
flask
pandas
numpy
opencv-python-headless
matplotlib
pillow
gunicorn
